Refactor auth token verification to a new location and update import paths in REST API routes
